Some modifications
Support check_mode
Added default protocol for url
Changed default variable for url
Added sanity to error messages
Language fixes

+++ lib/ansible/modules/web_infrastructure/sentry_project.py
@@ -24,7 +24,7 @@ version_added: "2.8"
 options:
   api_token:
     description:
-      - API token generated in Sentry. The token has to have an acess for manipulating projects.
+      - API token generated in Sentry. The token sufficient access to manipulate projects.
     required: true
 
   organization:
@@ -39,7 +39,7 @@ options:
 
   project_slug:
     description:
-      - Slug name. Slug name must be uniq (even after project deletion you can't create a project with same slug name).
+      - Slug name. Slug name must be unique (even after project deletion you can't create a project with same slug name).
     required: true
 
   state:
@@ -50,13 +50,15 @@ options:
 
   team:
     description:
-      - A team you are in and which has an access for project manipulating.
+      - The name of the team to create a new project for.
 
   url:
     description:
-      - Sentry URL with scheme.
-    required: true
+      - Sentry URL.
+    required: false
+    default: sentry.io
 
+notes: "Module supports check_mode, but it can't provide 100 percent gurantee what specified slug hasn't been used before." 
 author: "Mikhail Naletov (@okgolove)"
 '''
 
@@ -69,7 +71,7 @@ EXAMPLES = '''
     project_slug: backend
     state: present
     team: senior
-    url: https://sentry.example.com
+    url: sentry.example.com
 '''
 
 RETURN = ''' # '''
@@ -78,7 +80,7 @@ import json
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.urls import fetch_url
 from ansible.module_utils._text import to_native
-
+from ansible.module_utils.six.moves.urllib.parse import urlparse
 
 def main():
     arg_spec = dict(
@@ -88,10 +90,10 @@ def main():
         project_slug=dict(required=True),
         state=dict(required=True, choices=['present', 'absent']),
         team=dict(required=True),
-        url=dict(required=True)
+        url=dict(default="sentry.io")
     )
 
-    module = AnsibleModule(argument_spec=arg_spec, supports_check_mode=False)
+    module = AnsibleModule(argument_spec=arg_spec, supports_check_mode=True)
 
     api_token = module.params['api_token']
     organization = module.params['organization']
@@ -99,7 +101,10 @@ def main():
     project_slug = module.params['project_slug']
     state = module.params['state']
     team = module.params['team']
-    url = module.params['url']
+    url = urlparse(module.params['url'])
+    
+    if not url.scheme:
+      url = "https://{0}".format(url.path)
 
     def is_project_exists(project_name, organization, team, api_token):
         response, info = fetch_url(
@@ -108,14 +113,15 @@ def main():
             headers={"Authorization": "Bearer {0}".format(api_token)},
             method="GET")
         if info["status"] != 200:
-            module.fail_json(msg=info["msg"])
+            module.fail_json(msg="Unable to obtain a project list, response status: {0}, message: {1}".format(
+                  info["status"], info["msg"]))
         body = json.loads(response.read())
         return bool([x for x in body if x["name"] == project_name])
 
     if state == 'present':
         if is_project_exists(project_name, organization, team, api_token):
             module.exit_json(changed=False)
-        else:
+        if not module.check_mode:
             response, info = fetch_url(
                 module, "{url}/api/0/teams/{organization}/{team}/projects/".format(
                     url=url, organization=organization, team=team
@@ -130,22 +136,25 @@ def main():
                 }),
                 method="POST")
             if info["status"] != 201:
-                module.fail_json(msg=info["msg"])
-            module.exit_json(changed=True)
+                module.fail_json(msg="Unable to create the project, response status: {0}, message: {1}".format(
+                  info["status"], info["msg"]))
+        module.exit_json(changed=True)
 
     if state == 'absent':
         if is_project_exists(project_name, organization, team, api_token):
-            response, info = fetch_url(
-                module, "{url}/api/0/projects/{organization}/{project_slug}/".format(
-                    url=url, organization=organization,
-                    project_slug=project_slug),
-                headers={
-                    "Authorization": "Bearer {0}".format(api_token),
-                    "Content-Type": "application/json"
-                },
-                method="DELETE")
-            if info["status"] != 204:
-                module.fail_json(msg=info["msg"])
+            if not module.check_mode:
+                response, info = fetch_url(
+                    module, "{url}/api/0/projects/{organization}/{project_slug}/".format(
+                        url=url, organization=organization,
+                        project_slug=project_slug),
+                    headers={
+                        "Authorization": "Bearer {0}".format(api_token),
+                        "Content-Type": "application/json"
+                    },
+                    method="DELETE")
+                if info["status"] != 204:
+                    module.fail_json(msg="Unable to delete the project, response status: {0}, message: {1}".format(
+                    info["status"], info["msg"]))
             module.exit_json(changed=True)
         else:
             module.exit_json(changed=False)
